performed eda
